Replace MovieDetailsLoader protocol with MovieDetailsService

+++ MovieDemo/MovieDemo/Protocols/PersonDetailsService.swift
@@ -9,4 +9,4 @@
 import Foundation
 import Combine
 
-typealias PersonDetailsService = (Int) -> AnyPublisher<Person, Error>
+typealias PersonDetailsService = () -> AnyPublisher<Person, Error>
